config.py
import os
from dotenv import load_dotenv

load_dotenv()

ENV = os.getenv("FLASK_ENV")
DEBUG = ENV == "development"

SECRET_KEY = os.getenv("SECRET_KEY")
GOOGLE_CLIENT_ID = os.getenv("GOOGLE_CLIENT_ID")

SQLALCHEMY_DATABASE_URI = 'sqlite:///app.db'
SQLALCHEMY_TRACK_MODIFICATIONS = False
SQLALCHEMY_ECHO = False
