requirements.txt
alembic==1.7.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
